requirements.txt
pytest~=8.3.4
selene~=2.0.0rc9
selenium~=4.28.1
allure-pytest
allure-python-commons~=2.13.5
python-dotenv